fix: Show entire list of available branches
Updated the 'updateBranches' function in 'WorkflowRunListLoader.kt'. Instead of making a single request to get the branches, we are now clearing the "repoBranches" list and then making a paginated request to ensure we get all branches from the repository. Revised the function to iterate and make additional requests as long as 'response.hasNext' returns true, after which it updates 'repoBranches' with the newly fetched data. The bug was noticed due to incomplete display of branches in larger repositories.