fix: Update workflow run list on filter changes
The 'loadMore' function now takes true as argument to force update the workflow run list. This change was required as previously, the workflow run list did not reflect changes instantly when filters were updated. Now, with the force update in place, the run list will be correctly reloaded whenever there is a change in filters. And also in the case that the "Refresh" action will be clicked.