Missing branches and missing update on branch filer changes (#92)
* fix: Show entire list of available branches

Updated the 'updateBranches' function in 'WorkflowRunListLoader.kt'. Instead of making a single request to get the branches, we are now clearing the "repoBranches" list and then making a paginated request to ensure we get all branches from the repository. Revised the function to iterate and make additional requests as long as 'response.hasNext' returns true, after which it updates 'repoBranches' with the newly fetched data. The bug was noticed due to incomplete display of branches in larger repositories.

* fix: Update workflow run list on filter changes

The 'loadMore' function now takes true as argument to force update the workflow run list. This change was required as previously, the workflow run list did not reflect changes instantly when filters were updated. Now, with the force update in place, the run list will be correctly reloaded whenever there is a change in filters. And also in the case that the "Refresh" action will be clicked.

* fix: Load all available collaborators via pagination

Modified the updateCollaborators function in WorkflowRunListLoader to resolve an issue where not all collaborators were being loaded. The code now implements pagination to ensure all collaborators are fetched from the GitHub API.

* refact: Fetch data by using the nextLink from request response

The commit implements a change where the data fetching method in workflow run list loader now uses the `nextLink` from the previous request’s response instead of incrementing the page number after every request. This approach ensures data is retrieved effectively, as it correctly selects the continuation page from the previous call. This includes updates to how information for collaborators and branches is retrieved.